+++ TeamCode/src/main/java/org/betastudio/ftc/selection/SelectPackage.java
@@ -1,4 +1,4 @@
-package org.betastudio.ftc.util.selection;
+package org.betastudio.ftc.selection;
 
 import android.util.Range;
