+++ TeamCode/src/main/java/org/betastudio/ftc/time/Timestamp.java
@@ -1,4 +1,4 @@
-package org.betastudio.ftc.util.time;
+package org.betastudio.ftc.time;
 
 import androidx.annotation.NonNull;
