+++ TeamCode/src/main/java/org/betastudio/ftc/ui/log/FtcLogElement.java
@@ -1,7 +1,7 @@
 package org.betastudio.ftc.ui.log;
 
 import org.betastudio.ftc.util.message.LogMessage;
-import org.betastudio.ftc.util.time.Timestamp;
+import org.betastudio.ftc.time.Timestamp;
 
 public interface FtcLogElement {
 	Timestamp getTimestamp();
