+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/FtcLogFilesInspect.java
@@ -14,8 +14,8 @@ import org.betastudio.ftc.ui.telemetry.TelemetryElement;
 import org.betastudio.ftc.ui.telemetry.TelemetryLine;
 import org.betastudio.ftc.util.ButtonProcessor;
 import org.betastudio.ftc.util.message.TelemetryMsg;
-import org.betastudio.ftc.util.selection.SelectElement;
-import org.betastudio.ftc.util.selection.SelectPackage;
+import org.betastudio.ftc.selection.SelectElement;
+import org.betastudio.ftc.selection.SelectPackage;
 import org.firstinspires.ftc.teamcode.ButtonConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
