+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/ViewModeConvertor.java
@@ -7,8 +7,8 @@ import org.betastudio.ftc.ui.client.implementation.BaseMapClient;
 import org.betastudio.ftc.ui.client.ClientViewMode;
 import org.betastudio.ftc.ui.client.UpdateConfig;
 import org.betastudio.ftc.util.ButtonProcessor;
-import org.betastudio.ftc.util.selection.SelectElement;
-import org.betastudio.ftc.util.selection.SelectPackage;
+import org.betastudio.ftc.selection.SelectElement;
+import org.betastudio.ftc.selection.SelectPackage;
 import org.firstinspires.ftc.teamcode.ButtonConfig;
 
 import java.util.concurrent.atomic.AtomicReference;
