+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/tests/SelectPackageTest.java
@@ -7,8 +7,8 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import org.betastudio.ftc.ui.client.implementation.BaseMapClient;
 import org.betastudio.ftc.ui.client.UpdateConfig;
 import org.betastudio.ftc.util.ButtonProcessor;
-import org.betastudio.ftc.util.selection.SelectElement;
-import org.betastudio.ftc.util.selection.SelectPackage;
+import org.betastudio.ftc.selection.SelectElement;
+import org.betastudio.ftc.selection.SelectPackage;
 import org.firstinspires.ftc.teamcode.ButtonConfig;
 import org.firstinspires.ftc.teamcode.Global;
 import org.firstinspires.ftc.teamcode.RunMode;
